+++ CHANGELOG.md
@@ -5,6 +5,7 @@ a breaking change. Thanks to [José](https://github.com/nwparker) for his contri
 * Updated the Android example to AndroidX.
 * Added extra debug logging for language code.
 * Moved classes AlertStyleWidget, UpgradeBase, and UpgradeCard into their own Dart files.
+* Added Cupertino style alert test.
 
 ## 2.7.3